zp decim to int fixed

+++ src/billing/api/zarinpal.py
@@ -172,7 +172,7 @@ class ZPPayment(APIView):
         
         # Convert to Rials (Zarinpal uses Rials, multiply by 10 if you're storing in Toman)
         # Assuming amount is in Toman, convert to Rials
-        amount_rials = int(float(amount) * 10)
+        amount_rials = amount * 10
 
         # Create Payment record
         payment = PaymentModel.objects.create(
@@ -289,7 +289,7 @@ class ZPVerify(APIView):
             return redirect(f'https://app.pilito.com/dashboard/payment/success?payment_id={payment.ref_id}')
 
         # Verify with Zarinpal
-        amount_rials = int(float(payment.amount) * 10)
+        amount_rials = payment.amount * 10
         verify_data = {
             "MerchantID": settings.ZARRINPAL_MERCHANT_ID,
             "Amount": amount_rials,
